refactor(generateSlug): Extract duplicate code into a function.
